refactor(audio): extract FloatRingBuffer from main.cpp
