+++ Resona.vcxproj.filters
@@ -47,6 +47,9 @@
     <ClInclude Include="Audio\AudioDeviceManager.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="Audio\FloatRingBuffer.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="linear.comp">
